requirements.txt
arrow==1.3.0
astroid==3.3.8
alembic==1.17.1 
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cryptography==46.0.3
dill==0.3.9
Flask==3.1.0
flask-cors==6.0.1
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-migrate==4.1.0
Flask-JWT-Extended==4.7.1
idna==3.10
iniconfig==2.0.0
isort==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.10 
Markdown==3.7
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pydocstyle==6.3.0
pylint==3.3.4
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.3
six==1.17.0
snowballstemmer==2.2.0
soupsieve==2.6
tomlkit==0.13.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
